+++ front.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * COPS (Calibre OPDS PHP Server) front-end controller (dev only)
+ * COPS (Calibre OPDS PHP Server) front-end controller (dev only) @todo
  * with $config['cops_use_route_urls'] = '1' and no PHP script in URL
  *
  * @license    GPL 2 (http://www.gnu.org/licenses/gpl.html)
@@ -12,6 +12,10 @@ use SebLucas\Cops\Framework;
 
 require_once __DIR__ . '/config.php';
 
-//$request = Framework::getRequest();
-// @todo route to the the right endpoint if needed
+$request = Framework::getRequest();
+
+// @todo route to the right handler if needed
+$name = $request->getHandler('index');
+$handler = Framework::getHandler($name);
+//$handler->handle($request);
 var_dump($_SERVER);