+++ getJSON.php
@@ -14,7 +14,7 @@ require_once __DIR__ . '/config.php';
 
 // @todo handle 'json' routes correctly - see util.js
 //$request = Framework::getRequest('json');
-$request = Framework::getRequest();
+$request = Framework::getRequest('index');
 
 $handler = Framework::getHandler('json');
 $handler->handle($request);