+++ test/OPDSTest.php
@@ -35,7 +35,7 @@ class OpdsTest extends TestCase
         Config::set('calibre_directory', __DIR__ . "/BaseWithSomeBooks/");
         Database::clearDb();
         // try out route urls
-        Config::set('use_route_urls', true);
+        //Config::set('use_route_urls', true);
     }
 
     public static function tearDownAfterClass(): void
@@ -46,7 +46,7 @@ class OpdsTest extends TestCase
             return;
         }
         unlink(self::TEST_FEED);
-        Config::set('use_route_urls', null);
+        //Config::set('use_route_urls', null);
     }
 
     /**